tools/device_listener/requirements.txt
requests>=2.31.0
pyserial>=3.5
cryptography>=41.0.0
pyudev>=0.24.0
